Only delete data on plugin uninstall if user has opted-in for data deletion in EDD main settings. Closes #21

+++ edd-product-updates.php
@@ -150,22 +150,28 @@ register_deactivation_hook( __FILE__, 'edd_pup_delete_cron_schedule' );
  * @since 0.9.2
  */
 function edd_pup_uninstall(){
-    global $wpdb;
     
-    //Remove our table (if it exists)
-    $wpdb->query("DROP TABLE IF EXISTS $wpdb->edd_pup_queue");
+    global $edd_options;
     
-    //Remove all email posts
-    $wpdb->query("DELETE FROM $wpdb->posts WHERE post_type = 'edd_pup_email'");
-    
-    //Remove all custom metadata from postmeta table
-    $wpdb->query("DELETE FROM $wpdb->postmeta WHERE meta_key IN ( '_edd_pup_from_email' , '_edd_pup_from_name' , '_edd_pup_message' , '_edd_pup_recipients' , '_edd_pup_subject' , '_edd_pup_updated_products' )");
-    
-    //Remove all payment notes in customer payment histories
-    $wpdb->query("DELETE FROM $wpdb->comments WHERE comment_type = 'edd_payment_note' AND comment_content LIKE '%Sent product update%'");
-             
-    //Remove the version option
-    delete_option( 'edd_pup_version' );
+    if ( isset( $edd_options['uninstall_on_delete'] ) ) {
+	    
+	    global $wpdb;
+	    
+	    //Remove custom database table
+	    $wpdb->query("DROP TABLE IF EXISTS $wpdb->edd_pup_queue");
+	    
+	    //Remove all email posts
+	    $wpdb->query("DELETE FROM $wpdb->posts WHERE post_type = 'edd_pup_email'");
+	    
+	    //Remove all custom metadata from postmeta table
+	    $wpdb->query("DELETE FROM $wpdb->postmeta WHERE meta_key IN ( '_edd_pup_from_email' , '_edd_pup_from_name' , '_edd_pup_message' , '_edd_pup_recipients' , '_edd_pup_subject' , '_edd_pup_updated_products' )");
+	    
+	    //Remove all payment notes in customer payment histories
+	    $wpdb->query("DELETE FROM $wpdb->comments WHERE comment_type = 'edd_payment_note' AND comment_content LIKE '%Sent product update%'");
+	             
+	    //Remove the version option
+	    delete_option( 'edd_pup_version' );
+    } 
  
     //Remove any leftover transients
 	delete_transient( 'edd_pup_all_customers' );
